register promotion: add promoteToRegistersAtDepth
This function attempts register promotion below the given number of
schedule dimensions, splitting the bands to create a proper scope in the
tree if necessary.  It will be used in an upcoming commit.

I did consider having promotion depth relative to thread mapping, but
there is a situation I did not want to handle. Consider a schedule tree
with a sequence node after a band with 2 members. The first child has
two 2-member bands, with the innermost one mapped to threads. The
second child has only one 2-member band, which mapped to threads.
Relative depth "-4" would make me leave both children and perform
promotion _both_ above and below the common ancestor band. Does not
really make sense. Certainly, we can define a rule to handle such
situations, like the deepest common depth, but we want to go in a
different direction (band-local decisions in the tuner) anyway.